Changed recordscreen to use avconv instead of ffmpeg.
This is not a political change at all.  I'm staying out of the whole
ffmpeg split.  This is purely because I use Ubuntu, and I'm keeping
this script updated to work in that environment.
If anyone wishes me to maintain compatability with ffmpeg as well,
just let me know and offer me a bit of your time for testing.